+++ index.html
@@ -54,6 +54,6 @@
   <footer>
     <p>Copyright - Fatema Nazari , 2023</p>
   </footer>
-  <script src="index.js"></script>
+  <script src="index.js" type="module"></script>
 </body>
 </html>
